refactor(checker): split declaration_emit symbol-access helpers into own module
`declaration_emit.rs` reached 2024 lines, 24 over the 2000-line arch cap.
Extract the four `pub(crate)` symbol-accessibility helpers
(`get_symbol_from_any_binder`, `symbol_reachable_via_local_imports`,
`local_value_name_resolves_to`, `module_specifier_for_file`) into
`declaration_emit_symbol_access.rs` — a new sibling module under
`variable_helpers`. No semantic change; all callers remain valid because
`pub(crate)` visibility is crate-wide.

AgentName: M1FableArchitect
