+++ crates/tsz-checker/src/state/variable_checking/variable_helpers/mod.rs
@@ -3,3 +3,4 @@
 
 mod core;
 mod declaration_emit;
+mod declaration_emit_symbol_access;
